requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Flask-Bcrypt
Flask-SQLAlchemy
PyMySQL
